fix(docs): Improve alt texts of badge links
Yes the crates.io badge image displays as `[crates.io|$version]`,
where $version is the current crate version, however that doesn't
mean that `Crates.io Version` is a good alt text for the image.

The alt text of the image becomes the link label so it should describe
where the link takes you (as opposed to what the image displayed)
(besides "Version" isn't a helpful description of the image anyway).